Fix compile error after change to use C++23
list of compile error

[   45s] /home/abuild/rpmbuild/BUILD/dali2-2.5.17/dali/internal/event/actors/actor-sizer.cpp: In member function 'void Dali::Internal::ActorSizer::NegotiateDimension(Dali::Dimension::Type, const Dali::Vector2&, ActorDimensionStack&)':
[   45s] /home/abuild/rpmbuild/BUILD/dali2-2.5.17/dali/internal/event/actors/actor-sizer.cpp:793:21: error: C++20 says that these are ambiguous, even though the second is reversed: [-Werror]
[   45s]   793 |       if(element == searchActor)
[   45s]       |                     ^~~~~~~~~~~
[   45s] In file included from /home/abuild/rpmbuild/BUILD/dali2-2.5.17/dali/internal/event/actors/actor-sizer.cpp:18:
[   45s] /home/abuild/rpmbuild/BUILD/dali2-2.5.17/dali/internal/event/actors/actor-sizer.h:69:10: note: candidate 1: 'bool Dali::Internal::ActorSizer::ActorDimensionPair::operator==(const Dali::Internal::ActorSizer::ActorDimensionPair&)'
[   45s]    69 |     bool operator==(const ActorDimensionPair& rhs)
[   45s]       |          ^~~~~~~~
[   45s] /home/abuild/rpmbuild/BUILD/dali2-2.5.17/dali/internal/event/actors/actor-sizer.h:69:10: note: candidate 2: 'bool Dali::Internal::ActorSizer::ActorDimensionPair::operator==(const Dali::Internal::ActorSizer::ActorDimensionPair&)' (reversed)
[   45s] /home/abuild/rpmbuild/BUILD/dali2-2.5.17/dali/internal/event/actors/actor-sizer.h:69:10: note: try making the operator a 'const' member function

/home/hpichu/work/dali-core/dali/public-api/object/base-handle.h: In instantiation of ‘bool Dali::operator==(const Dali::BaseHandle&, const T&) [with T = Dali::WeakHandle<Dali::Actor>]’:
/home/hpichu/work/dali-core/dali/internal/event/render-tasks/render-task-impl.cpp:228:28:   required from here
/home/hpichu/work/dali-core/dali/public-api/object/base-handle.h:289:17: error: invalid ‘static_cast’ from type ‘const Dali::WeakHandle<Dali::Actor>’ to type ‘const Dali::BaseHandle&’
  289 |   return lhs == static_cast<const BaseHandle&>(rhs);
      |                 ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

/home/hpichu/work/dali-core/dali/internal/update/gestures/scene-graph-pan-gesture.cpp:91:5: error: ‘++’ expression of ‘volatile’-qualified type is deprecated [-Werror=volatile]
   91 |   ++mWritePosition;

/home/hpichu/work/dali-core/dali/internal/update/manager/update-proxy-property-modifier.h:150:47: error: no matching function for call to ‘Dali::Internal::UpdateProxy::PropertyModifier::NodePropertyPair::NodePropertyPair(<brace-enclosed initializer list>)’
  150 |     NodePropertyPair pair{&node, &propertyBase};

/home/hpichu/work/dali-core/dali/public-api/object/property-key.cpp:59:25: error: ambiguous overload for ‘operator==’ (operand types are ‘Dali::String’ and ‘Dali::StringView’)
   59 |     result = (stringKey == Dali::StringView(rhs.Data(), rhs.Size()));
      |               ~~~~~~~~~ ^~       ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
      |               |                  |
      |               Dali::String       Dali::StringView

/home/hpichu/work/dali-core/dali/devel-api/threading/conditional-wait.cpp:118:13: error: ‘++’ expression of ‘volatile’-qualified type is deprecated [-Werror=volatile]
  118 |   ++(mImpl->count);

/home/hpichu/work/dali-core/automated-tests/src/dali/utc-Dali-ConditionalWait.cpp:166:7: error: ‘--’ expression of ‘volatile’-qualified type is deprecated [-Werror=volatile]
  166 |     --gNotifyCount;

In file included from /home/hpichu/work/dali-core/automated-tests/src/dali/utc-Dali-TypeRegistry.cpp:19:
/home/hpichu/work/dali-core/automated-tests/src/dali/../../../dali/devel-api/object/type-registry-helper.h:39:67: error: arithmetic between different enumeration types ‘TypeRegistryTest::TestHandle::Property::<unnamed enum>’ and ‘TypeRegistryTest::TestHandle::PropertyRange’ is deprecated [-Werror=deprecated-enum-enum-conversion]
   39 |   static_assert((handleNamespace::handleType::Property::enumIndex - handleNamespace::handleType::PROPERTY_START_INDEX) == count);

[   71s] /home/abuild/rpmbuild/BUILD/dali2-2.5.17/automated-tests/src/dali/dali-test-suite-utils/test-graphics-reflection.cpp:275:10: error: use of deleted function 'std::__cxx11::basic_string<_CharT, _Traits, _Alloc>::basic_string(std::nullptr_t) [with _CharT = char; _Traits = std::char_traits<char>; _Alloc = std::allocator<char>; std::nullptr_t = std::nullptr_t]'
[   71s]   275 |   return 0u;

+++ dali/devel-api/object/property-helper-devel.h
@@ -29,11 +29,11 @@
  */
 #define DALI_DEVEL_PROPERTY_REGISTRATION_INTERNAL(count, typeRegistrationObject, objectNamespace, objectType, develNamespace, text, valueType, enumIndex)                                                                           \
   Dali::PropertyRegistration DALI_TOKEN_PASTE(property, count)(typeRegistrationObject, text, objectNamespace::develNamespace ::Property::enumIndex, Dali::Property::valueType, &objectType::SetProperty, &objectType::GetProperty); \
-  static_assert((objectNamespace::develNamespace ::Property::enumIndex - objectNamespace::objectType::PROPERTY_START_INDEX) == count);
+  static_assert((static_cast<int>(objectNamespace::develNamespace ::Property::enumIndex) - static_cast<int>(objectNamespace::objectType::PROPERTY_START_INDEX)) == count);
 
 #define DALI_DEVEL_PROPERTY_REGISTRATION_INTERNAL_READ_ONLY(count, typeRegistrationObject, objectNamespace, objectType, develNamespace, text, valueType, enumIndex)                                             \
   Dali::PropertyRegistration DALI_TOKEN_PASTE(property, count)(typeRegistrationObject, text, objectNamespace::develNamespace ::Property::enumIndex, Dali::Property::valueType, NULL, &objectType::GetProperty); \
-  static_assert((objectNamespace::develNamespace ::Property::enumIndex - objectNamespace::objectType::PROPERTY_START_INDEX) == count);
+  static_assert((static_cast<int>(objectNamespace::develNamespace ::Property::enumIndex) - static_cast<int>(objectNamespace::objectType::PROPERTY_START_INDEX)) == count);
 
 /**
  * @brief These macros are used to define properties for implementations of CustomActor.
